# docs/04-api-reference/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NextCRM API",
    "description": "NextCRM is a comprehensive Customer Relationship Management system with multi-tenant architecture, subscription billing, and project management capabilities. This API provides programmatic access to all CRM features including accounts, contacts, leads, opportunities, invoices, projects, and document management.",
    "version": "3.0.0-alpha",
    "contact": {
      "name": "NextCRM Support",
      "url": "https://github.com/DrivenIdeaLab/nextcrm-app",
      "email": "support@nextcrm.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.nextcrm.io/api",
      "description": "Production API Server"
    },
    {
      "url": "https://demo.nextcrm.io/api",
      "description": "Demo API Server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/crm/contacts/{contactId}": {
      "get": {
        "summary": "Get contact by ID",
        "description": "Retrieve a single contact by its ID",
        "tags": ["CRM - Contacts"],
        "operationId": "getContactById",
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Update a contact",
        "description": "Update contact information",
        "tags": ["CRM - Contacts"],
        "operationId": "updateContact",
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Delete a contact",
        "description": "Delete a contact from the system",
        "tags": ["CRM - Contacts"],
        "operationId": "deleteContact",
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/projects/tasks/create-task": {
      "post": {
        "summary": "Create a new task",
        "description": "Create a new task in a project section",
        "tags": ["Projects"],
        "operationId": "createTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/organization": {
      "get": {
        "summary": "Get organization details",
        "description": "Retrieve the current user's organization information including owner and member list",
        "tags": ["Organization"],
        "operationId": "getOrganization",
        "responses": {
          "200": {
            "description": "Organization details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new organization",
        "description": "Create a new organization. User must not already belong to an organization.",
        "tags": ["Organization"],
        "operationId": "createOrganization",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update organization",
        "description": "Update organization details. Only organization owner or admin can update.",
        "tags": ["Organization"],
        "operationId": "updateOrganization",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/organization/members": {
      "get": {
        "summary": "List organization members",
        "description": "Get a list of all members in the organization",
        "tags": ["Organization"],
        "operationId": "listOrganizationMembers",
        "responses": {
          "200": {
            "description": "List of organization members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/billing/subscription": {
      "get": {
        "summary": "Get subscription details",
        "description": "Retrieve organization subscription and payment history. Only accessible by organization owner.",
        "tags": ["Billing"],
        "operationId": "getSubscription",
        "responses": {
          "200": {
            "description": "Subscription details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionDetails"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/OwnerOnly"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from NextAuth authentication"
      }
    },
    "schemas": {
...
      "Contact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "first_name": {
            "type": "string",
            "nullable": true
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "personal_email": {
            "type": "string",
            "nullable": true
          },
          "office_phone": {
            "type": "string",
            "nullable": true
          },
          "mobile_phone": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "account": {
            "type": "string",
            "nullable": true
          },
          "assigned_to": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "boolean",
            "default": true
          },
          "type": {
            "type": "string",
            "enum": ["Customer", "Partner", "Vendor", "Prospect"],
            "default": "Customer"
          },
          "social_twitter": {
            "type": "string",
            "nullable": true
          },
          "social_facebook": {
            "type": "string",
            "nullable": true
          },
          "social_linkedin": {
            "type": "string",
            "nullable": true
          },
          "social_skype": {
            "type": "string",
            "nullable": true
          },
          "social_instagram": {
            "type": "string",
            "nullable": true
          },
          "social_youtube": {
            "type": "string",
            "nullable": true
          },
          "social_tiktok": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": ["id", "organizationId", "last_name"]
      },
...
      "ContactUpdateRequest": {
        "type": "object",
        "required": ["id", "last_name"],
        "properties": {
          "id": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "personal_email": {
            "type": "string"
          },
          "office_phone": {
            "type": "string"
          },
          "mobile_phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "account": {
            "type": "string"
          },
          "assigned_to": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": ["Customer", "Partner", "Vendor", "Prospect"]
          }
        }
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "section": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "type": "integer"
          },
          "priority": {
            "type": "string"
          },
          "dueDateAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "user": {
            "type": "string",
            "nullable": true
          },
          "taskStatus": {
            "type": "string",
            "enum": ["ACTIVE", "PENDING", "COMPLETE"],
            "default": "ACTIVE"
          },
          "likes": {
            "type": "integer",
            "default": 0
          },
          "tags": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TaskCreateRequest": {
        "type": "object",
        "required": ["title", "section", "position", "priority"],
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          },
          "priority": {
            "type": "string"
          },
          "dueDateAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "string"
          }
        }
      },
...
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "plan": {
            "type": "string",
            "enum": ["FREE", "PRO", "ENTERPRISE"]
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "SUSPENDED", "CANCELLED"]
          },
          "ownerId": {
            "type": "string"
          },
          "stripeCustomerId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "owner": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          },
          "users": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "OrganizationCreateRequest": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Organization name"
          },
          "slug": {
            "type": "string",
            "description": "Unique URL-friendly identifier",
            "pattern": "^[a-z0-9-]+$"
          }
        }
      },
      "OrganizationUpdateRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "userStatus": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE", "PENDING"]
          },
          "userLanguage": {
            "type": "string",
            "enum": ["en", "de", "cz", "uk"]
          },
          "is_admin": {
            "type": "boolean"
          },
          "organizationId": {
            "type": "string",
            "nullable": true
          },
          "organization_role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "MEMBER", "VIEWER"],
            "nullable": true
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_on": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or session invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Unauthenticated",
              "message": "Valid authentication token is required",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions to access this resource",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Forbidden",
              "message": "You do not have permission to access this resource",
              "code": "FORBIDDEN"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource was not found or user does not have access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Not Found",
              "message": "Account not found or unauthorized",
              "code": "NOT_FOUND"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request parameters or malformed request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Bad Request",
              "message": "Name and slug are required",
              "code": "INVALID_REQUEST"
            }
          }
        }
      },
...
    }
  }
}